refactor: enum 디렉토리 분리
enumeration 패키지 하위에 두어 구분하도록 하였습니다.
