+++ src/main/java/lotto/domain/enumeration/BallType.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.enumeration;
 
 public enum BallType {
     NORMAL,
