+++ src/test/java/lotto/domain/LottoTicketTest.java
@@ -3,6 +3,7 @@ package lotto.domain;
 import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
+import lotto.domain.enumeration.BallType;
 import lotto.domain.generator.AutoLottoNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
